moved the profile schema into its own directory
