+++ server/app.js
@@ -16,51 +16,6 @@ mongoose.connect(process.env.DATABASE_CONNECT, {
   useUnifiedTopology: true,
 });
 
-// mongoose schema
-const profileSchema = {
-  userID:{
-    type: mongoose.Types.ObjectId
-  },
-  firstName:{
-    type: String,
-    required: true
-  },
-  lastName:{
-    type: String,
-    required: true
-  },
-  gender:{
-    type: String,
-    required: true
-  },
-  birthDate:{
-    type: Date,
-    required: true
-  },
-  email:{
-    type: String,
-    required: true
-  },
-  phone:{
-    type: Number,
-    required: true
-  },
-  address:{
-    type: String,
-    required: true
-  },
-  description:{
-    type: String,
-    required: true
-  },
-  availability:{
-    type: Boolean,
-    required: true
-  }
-}
-
-const Profile = mongoose.model('Profile', profileSchema);
-
 var app = express();
 
 app.use(logger("dev"));
